split spinlocks to separate files
